In Progress: Implementing Day Insertion API
Decent progress in handling Timezones. If a user puts time into a specific day, then it should always be bound to that specific day.
For example, if a user submits time to 1/31/2020, no matter what timezone they view the data in, it will always be 1/31/2020.

Implemented that by converting local time into a UTC Date object with the same day.

Posting a new day is done,  but needs testing. Probably the most complex part of this API.

Created a bisectLeft() that is very similar to Python's bisect_left() module for implementing binary search and made it so that my date utility functions zero out time.